requirements.txt
flask
flask-cors
pymongo
azure-cosmos
flask_pymongo
python-dotenv
gunicorn
azure-storage-blob
